Added GetUserAccountMoviePageUseCase
